XD-437 Removed package tangles identified by Sonar.
Changes included:
* Moved abstract Deployer to the stream package because this is the only place it is used.  We probably want to create a deploy package and move all deploy based classes there.
* Moved Base Definition to core because its used by ResourceDefinition in the core package, and the ResourceDefinition is used by the streams package.

* The package tangles between streams.dsl and stream.dsl.ast were fairly extensive.  Specifically around  StreamConfigParser having use of the nodes in the ast package and the Nodes having use in the dsl package.  The relationship was fairly tight, so I merged the 2 back together.  The only option that came to mind is to create a interface for the StreamConfigParser, but I was concerned, that I would be creating a patch instead of a real fix.  We may want to revisit the package structure here.
