+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/stream/TriggerDefinition.java
@@ -12,6 +12,8 @@
  */
 package org.springframework.xd.dirt.stream;
 
+import org.springframework.xd.dirt.core.BaseDefinition;
+
 
 import org.springframework.xd.dirt.core.BaseDefinition;
