+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/stream/dsl/LabelNode.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.xd.dirt.stream.dsl.ast;
+package org.springframework.xd.dirt.stream.dsl;
 
 /**
  * Represents a label attached to a module.
