+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/stream/dsl/ModuleReferenceNode.java
@@ -13,31 +13,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.xd.dirt.stream.dsl.ast;
+package org.springframework.xd.dirt.stream.dsl;
+
 
 /**
  * @author Andy Clement
  */
-public class ChannelNode extends AstNode {
+public class ModuleReferenceNode extends AstNode {
 
 	private final String streamName;
-	private final String channelName;
+	private final String labelOrModuleName;
+	
+	private String resolvedChannel;
 
-	public ChannelNode(String streamName, String channelName, int startpos, int endpos) {
+	public ModuleReferenceNode(String streamName, String moduleName, int startpos, int endpos) {
 		super(startpos,endpos);
 		this.streamName = streamName;
-		this.channelName = channelName;
+		this.labelOrModuleName = moduleName;
 	}
 
 	@Override
 	public String stringify(boolean includePositionalInfo) {
 		StringBuilder s = new StringBuilder();
 		s.append("(");
-		s.append(":");
 		if (streamName != null) {
 			s.append(streamName).append(".");
 		}
-		s.append(channelName);
+		s.append(labelOrModuleName);
+		if (resolvedChannel!=null) {
+			s.append("[[channel:").append(resolvedChannel).append("]]");
+		}
 		if (includePositionalInfo) {
 			s.append(":");
 			s.append(getStartPos()).append(">").append(getEndPos());
@@ -45,14 +50,13 @@ public class ChannelNode extends AstNode {
 		s.append(")");
 		return s.toString();
 	}
-
+	
 	public String toString() {
 		StringBuilder s = new StringBuilder();
-		s.append(":");
 		if (streamName != null) {
 			s.append(streamName).append(".");
 		}
-		s.append(channelName);
+		s.append(labelOrModuleName);
 		return s.toString();
 	}
 
@@ -63,12 +67,27 @@ public class ChannelNode extends AstNode {
 		return streamName;
 	}
 	
-	public String getChannelName() {
-		return channelName;
+	public String getModuleName() {
+		return labelOrModuleName;
 	}
 
-	public ChannelNode copyOf() {
-		return new ChannelNode(streamName,channelName,startpos,endpos);
+	public void resolve(StreamLookupEnvironment env) {
+		resolvedChannel = env.lookupChannelForLabelOrModule(streamName, labelOrModuleName);
+		if (streamName==null && resolvedChannel == null) {
+			// it is possible the singular name in labelOrModuleName is actually
+			// a stream reference
+			StreamNode sn = env.lookupStream(labelOrModuleName);
+			if (sn!=null) {
+				resolvedChannel = sn.getStreamName()+".0";
+			}
+		}
+	}
+
+	public ModuleReferenceNode copyOf() {
+		ModuleReferenceNode moduleReferenceNode =
+			new ModuleReferenceNode(streamName, labelOrModuleName, startpos, endpos);
+		moduleReferenceNode.resolvedChannel = resolvedChannel;
+		return moduleReferenceNode;
 	}
 
 }
