+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/stream/dsl/SourceChannelNode.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.xd.dirt.stream.dsl.ast;
+package org.springframework.xd.dirt.stream.dsl;
 
-import org.springframework.xd.dirt.stream.dsl.StreamLookupEnvironment;
 
 /**
  * @author Andy Clement
