+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/stream/dsl/StreamNode.java
@@ -13,13 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.xd.dirt.stream.dsl.ast;
+package org.springframework.xd.dirt.stream.dsl;
 
 import java.util.List;
 
-import org.springframework.xd.dirt.stream.dsl.DSLException;
-import org.springframework.xd.dirt.stream.dsl.StreamLookupEnvironment;
-import org.springframework.xd.dirt.stream.dsl.XDDSLMessages;
 
 /**
  * @author Andy Clement
