+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/stream/dsl/StreamsNode.java
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.xd.dirt.stream.dsl.ast;
+package org.springframework.xd.dirt.stream.dsl;
 
 import java.util.List;
 
-import org.springframework.xd.dirt.stream.dsl.StreamLookupEnvironment;
 
 /**
  * @author Andy Clement
